Introduce a cacheURLs array to prevent race conditions related to checking the on-disk image cache.
If a cache check is initiated, then canceled with cancelForDelegate: before a new cache check is started with a different URL but the same delegate, the ongoing disk cache check would still message the delegate despite being canceled by cancelForDelegate:. This is because it only checked to see if the delegate was in the cacheDelegates array; it had been removed by cancelForDelegate:, but added back by the new unrelated cache check. cacheURLs keeps track of which specific URLs are actually requested by the delegate. If the URL from a completed disk cache check does not match the delegate in cacheDelegates, the delegate is not messaged (since we know that request was canceled).